* Improve the code in vars.c * - Change 'flags' to 'pending' since that's all it is used for * - Change $symbolctl() to support both FLAGS and PENDING for the symbols * - Make get_string_var() return a const ptr so people don't change it * - Make make_string_var_bydata() take an (IrcVariable) since that's easier * Crank up the warning flags and start addressing things * - Add a test in configure for __attribute__((fallthrough)) * - This allows the use of FALLTHROUGH to shut up warnings. * - Improve a lot of const correctness issues * - Fix other bogons it pointed out (d'oh!)

+++ source/ssl.c
@@ -629,7 +629,7 @@ static char *get_bio_mem_string(BIO *mem)
 
 	bytes = BIO_get_mem_data(mem, &p);
 	
-	if (p && bytes >= 0 && bytes < SIZE_MAX)
+	if (p && bytes >= 0 && (unsigned long)bytes < SIZE_MAX)
 	{
 		char *result = new_malloc((size_t)bytes + 1);
 		memcpy(result, p, bytes);
